Fix: restore ability to load older profile versions
#### Brief overview of PR changes/additions
Ensures that the case where the profile to load is selected from the GUI
dialogue passes the particular file that the user can select via that form
is loaded rather than automatically loading the newest file that other uses
of the profile loading code refactored by #7395 uses.

#### Motivation for adding to Mudlet
Fixes a bug introduced by the PR 7395.

#### Other info (issues closed, discussion etc)
This is obviously something that was not spotted when the original PR was
"tested" - though to be honest it was so something that the creator of that
did seem to worry about:
> Unifying the duped logic was a bit intimidating seeing as most of those
lines haven't been touched in a decade or so, but I tried to keep actual
changes in what is being executed minimal. Though I did find a few
things that were being done in only one branch but not the other, which
seemed to me like they should have really been in both, and them not
being so was a bug.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>