Fix: reduce C++20 warnings
#### Brief overview of PR changes/additions
Rephrases some combinations of Qt `enum`s that whilst acceptable for C++17
upset the later one. For the non-key related ones which are provided as
`int`s arguments to a method this is sufficient; for the key ones however
it is most effective to resort to the form that takes a string to describe
the combination rather than to use the `enum`s that Qt already provides.
For Qt 6 there are some additional methods that overcome the
"incompatibilty" of combining such different `enum`s but those have not been
back ported to the Qt 5.15.8 I have - despite what https://bugreports.qt.io/browse/QTBUG-99948
says.

#### Motivation for adding to Mudlet
To eliminate the following type of warning when building with a C++20
compiler:
* warning: bitwise operation between different enumeration types
‘QFont::Weight’ and ‘QFont::StyleHint’ is deprecated [-Wdeprecated-enum-enum-conversion]
* warning: bitwise operation between different enumeration types
‘Qt::Modifier’ and ‘Qt::Key’ is deprecated [-Wdeprecated-enum-enum-conversion]
* warning: bitwise operation between different enumeration types
‘Qt::TextFlag’ and ‘Qt::AlignmentFlag’ is deprecated [-Wdeprecated-enum-enum-conversion]

#### Other info (issues closed, discussion etc)
The code in the `QFont` cases contains errors that have been copied from
bogus QFont creation code going back to the very first (well second) commit
in the git history. This contained calls of the form
`QFont font("Courier New", 10, QFont::Courier)`
however even the Qt 4.8 documentation does not list a constructor of that
form but instead has:
`QFont(const QString & family, int pointSize = -1, int weight = -1,
bool italic = false)` the third argument could possibly be `QFont::Normal`
(50) or `QFont::Bold` (75) however the value of `QFont::Courier` is `2`
but it is for a completely different purpose, that of the font matching
strategy ("the font matcher prefers fixed pitch fonts.") but that is not
something that can be set as an argument to the font constructor!

As it happens the combination of `QFont::Bold | QFont::Serif | QFont::PreferMatch | QFont::PreferAntialias`
that was being used numerically equals 75 + 2 + 32 + 128 = 227 - and the
scale that Qt actually uses only goes from 0 to 99!

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>