Infrastructure: use better/correct C++20 selection logic for qmake builds
#### Brief overview of PR changes/additions
Rewrite the part of the qmake project file so the logic and choices are
correct.

#### Motivation for adding to Mudlet
The previous logic was (incorrectly)
  IF Qt Major version is less than 5
  OR (if Qt Major version is less than 6 AND if Qt Minor version is less
than 12))
    THEN add `-std=c++20` to `QMAKE_CXXFLAGS`
    ELSE add `c++2a` to `CONFIG`

This is borked because we have already rejected Qt versions less than 5.14
so the logic will **always** end up in the **ELSE** case, and whilst Qt 5.15 is
documented accepting `CONFIG += c++2a` (and `c++2b`) the earliest Qt 6.x
version describes `c++2a` as an obsolete alias for `c++20` (in
https://doc.qt.io/qt-6.2/qmake-variable-reference.html).

#### Other info (issues closed, discussion etc)
This change makes the code match the comments!

Overall all the changes in this PR means that #7613 is not required after
all - at least for Mudlet's own code - though there are still some warnings
from the edbee-lib sub-module.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>