Fix: put an outline around command lines
#### Brief overview of PR changes/additions
Adds a frame around the `QPlainLineEdit` used for every command line.

#### Motivation for adding to Mudlet
Reported by the Discord user "cressie007" in the Mudlet Discord Guild
**Help** channel - in the thread "Issue I had after updating to 4.19,1",
was the fact that there was no visible demarcation between the input and
output areas in the main console. A closer examination showed me that it
was consistent across recent versions (4.18.5 at least to present) and
visible with the dark mode (which is based on a modified "Fusion" theme
that ALL builds of Mudlet will use on all OSes.) The reporter did note that
when they reverted to 4.17.1 they did not have this issue however I had
it to be present in both 4.17.0 and 4.17.2 release version.

#### Other info (issues closed, discussion etc)
The colour of the boarder/frame can be style specific but there is now
definitely something showing after this PR.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>