Improve: make exits to zero be stubs
#### Brief overview of PR changes/additions
*Extensive* change-set to allow the creation of stubs on special exits and
the addition of custom exit lines on those as well as "real" exits. UI
modified to support this (removal of dedicated "Stub exit" check-boxes on
"Room Exits" dialog, new status icon for stubs in exit room Id line-edits),
amendments to custom line dialogues to handle lines that do not go to a
specified room.

A fix to and enabling of map version 21 which eliminates the manipulations
to the map data so that the internal data structures - which have developed
over time - are stored in a compatible manner in older format. This largely
makes use of the "User Data" parts of the Map/Area/Room constructs. Whilst
developing this code I discovered that my previous work in this area was
leaving the modified User Data in place after the save had taking place and
with the separate development of the "background saves" this meant it was
being left in place even when the map format in use didn't not need it; to
fix this a revision has been made that removes those "system.fallbackXXXX"
user data items as soon as the corresponding user data item has been
written to file storage.

With the addition of the ability to have a custom exit line for stub exits
that they could, for instance point to a Map Label. There is also the
addition of a circle line around rooms that have a special exit which does
NOT have a custom exit line and if there are none of those a dotted circle
if there is a special stub exit without a custom line - this is so that
end-users can see that there are special exits/stubs in a room when there
is no other indications of one being there. Also for non-XY plane "normal"
exits and stubs previous Mudlet version would display triangle shapes on
the room square/circle to indicate their presence - these are now
suppressed if there is a custom exit line, this reduces the "stuff" on the
room if it is not needed.

#### Motivation for adding to Mudlet
Improved map creation end-user functionality as it allows better handling
of exits whose destination is not known. Also, since "Special" exit stubs
can theoretically be scripts it ought to be possible to handle exits that
do NOT go to a fixed location - although that will need mapper packages to
handle it.

#### Other info (issues closed, discussion etc)
The way the extra data is stored in the current map formats will enable
maps with these extra features to be opened in current Mudlet version
although they (custom lines on stubs and special exit stubs) will not be
visible in the mapper of such versions. With the enabling of binary map
format 21 maps saved with that format will not be open-able with Mudlet
versions prior to this PR being merged - but the default has not been
changed so it will not be used without deliberate action by the end-user.

Also added some debug messages for binary map file operations - to make it
easier to debug in the future when revising the map format.
`DEBUG_MAP_FILE_PROCESSING` is not defined to anything by default but
setting it to `1`, `2` or `3` will give increasingly more reports of the
current position in the file when reading and writing the file - which
helps to work out where things are failing should the two sets of numbers
get out of step.

Also added some "magic" - the ASCII characters "MudletMapFile" immediately
following the 4 bytes big-endian `int32_t` that contains the map file
version for version 21 (or later) map file format. This is to make the
detection of files as actually being binary Map Files for Mudlet even more
reliable than currently - and could be submitted to:
https://www.darwinsys.com/file/

Also remove the prefix from some automatic variables that start with the
ill-advised `_` character.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>