Revise: add format version handling & user-data fallbacks in JSON map file.
I initially included map symbol fonts in the code written - before I
realised I had already incorporated that data in the 1.000 format! Whilst
I removed the specific writing of that details into the JSON map file I
left the generic (and `static`) reader and writer methods in place should
they be needed in future.

Previously the JSON map format did not record the map zoom level for each
area - as that was introduced after the original JSON map handling was done
and I forgot to put it in. It is now recorded in the area user-data of JSON
map files but removed from the user data on reading and immediately from
the current profile data immediately after writing it to the JSON file.

Also I discovered that I had made a coding error in the prior JSON reading
code that meant colours with an alpha component were not being read
correctly - they lost the alpha and were drawn fully opaque whether it was
recorded as less than that level of opacity or not.

The JSON code has been revised to pass a version number (a decimal number
with 3 decimal places) throughout the reading and writing methods and code
put in to conditionally process the current fallback user data elements
that are needed but which will not be needed when the version value -
currently `1.000` is increased to `1.001` or more. This has not been
incremented so far.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>