Improve: revise 3Kingdoms and 3Scapes icons
#### Brief overview of PR changes/additions
Revised the icons used for 3Scapes and 3Kingdoms MUDS.

#### Motivation for adding to Mudlet
A Discord user (**cabbage3469**) pointed out that the logo for 3Kingdoms
was unreadable and upon checking it and comparing to how it appears on
their [website](http://www.3k.org/about3k.php) I concurred with their
assessment. This PR should improve things so that it is clear the MUDs that
this and the 3Scapes one refer to.

#### Other info (issues closed, discussion etc)
I took a large screen-shot of the logo (and the one for 3Scapes which
wasn't quite so bad but also did not conform to the 4:1 aspect ratio we
need) and then resized and produced a new icon preserving the proportions
of the origin website logo. This did however leave a large amount of space
to the right of the now approximately square logo on a coloured background
which I had to extend and also did not improve the readability of (either)
games' name in the logo. So I tried to find a font on my system that
matched the font used on the website in the flyer http://www.3k.org/images/flyers/flyer225.jpg
and used that to put the MUD name in white in the space to the right of the
logo.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>