Revise: reset mMaxVersion back down to 20
I've realised that moving the Map Labels to each area will make things
harder in the future for some improvements and I'll need to revert some
prior changes. As such format 21 is not as I'd wish it to be so I'll
disable it again. If anyone else HAS used it (though it is explicitly
marked as "experimental") they will need to temporarily get a copy of the
old code, build it, load the map and then resave it as a version 20 file
going forward.