Improve: add processor compilation details to getOS() output
#### Brief overview of PR changes/additions
Adds an addition value to those returned by the Lua API `getOS()` call to
include something about the processor the code was compilied for.

#### Motivation for adding to Mudlet
This is so that packagers that include Lua libraries can select the correct
one when the library is a binary one that is processor specific - e.g.
`arm` vs. `x86` when otherwise the library name would be the same for both.

#### Other info (issues closed, discussion etc)
The need for this came up in a discussion between myself and @jmckison
(@humera1295 on Discord) on 2025/02/02 15:19 UTC:
https://www.discord.com/channels/283581582550237184/283582439002210305/1335625345189019783

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>