Fix: keep CodeFactor happy by explicitly marking an unused argument
Whilst it is valid C++ code to not provide a name for an unused argument
in a method definition it confuses CodeFactor which takes it to be an old-
style C type "cast". By putting an argument and then wrapping it with the
QT `Q_UNUSED(`...`)` macro we avoid this confusion.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>