Fix: remove some more warnings both Qt version related and others
#### Brief overview of PR changes/additions
Replace usages of the `qAsConst(`...`)` wrapper with the STL
`std::as_const(`...`)` which has been recommended for Qt 6 as Qt remove
things that are now standard in the C++ standards.

#### Motivation for adding to Mudlet
Reduction of warnings when compiling with later Qt version.

#### Other info (issues closed, discussion etc)
This pair seems to have not been picked up by #7224 which already noted
that `qAsConst(`...`)` has been deprecated in Qt 6.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>