Improve: add Noto-Color-Emoji font support to FreeBSD builds
Extends the use of Noto-Color-Emoji to provide Emojis to the FreeBSD OS
as well as the GNU/Linux one.

I've recently discovered that FreeBSD now supports the display of colour
fonts (at least the type that this font is) so wanted them to be used.

I've renamed the Qt resource file that bundles the Noto-Color-Emoji font
to reflect the fact that it is now usable for more than just GNU/Linux.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>