Improve: provide "NewLoad/NotReset" flag to sysLoadEvent
#### Brief overview of PR changes/additions
Adds a boolean argument to the Mudlet `sysLoadEvent` which is `true` for
a profile being loaded and `false` after a use of `resetProfile()`.

#### Motivation for adding to Mudlet
Packages/scripts might need to behave differently if they get this event
when the profile is already in use as opposed to being freshly started.
For instance in the former case they may already be connected and active
on the Game Server or they may need to reacquire data/variables/tables in
a different manner.

#### Other info (issues closed, discussion etc)

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>