Test: provide a function to report working directory and library paths
#### Brief overview of PR changes/additions
Adds `applicationPaths()` function to Lua API which reports the Mudlet
application's idea of its current working directory and the location(s)
from which it would look to find binary libraries.

#### Motivation for adding to Mudlet
A user was reporting issues where Mudlet would load their first profile,
including loading all the Lua binary libraries (such as the Lua File
System) one and then fail to find it for a second profile, given that in
their case (Windows 64 bit) release version this would be from the same
directory as the Mudlet executable itself (the `./lfs.dll` entry reported)
in the error message in the main console. This function is intended to
enable the "current" working directory to be monitored - so that if it gets
changed it can be detected.

#### Other info (issues closed, discussion etc)
**This is a test/draft PR - it is not intended to be merged into the main
code-base.** The user concerned with whom I have been helping with this is
the Discord user `kyrus78`.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>