Infrastructure: investigate #7740
#### Brief overview of PR changes/additions
Adds a couple of debug output lines to report the Mudlet application's
idea of what it's "current working directory" is immediately prior to
running the `TLuaInterpreter::loadGlobal()` method.

#### Motivation for adding to Mudlet
This is designed to aide in the diagnosis of the issue #7740 which only
seems to manifest in Release and PTB builds of Mudlet on Windows.


#### Other info (issues closed, discussion etc)
I think this will be needed to be merging into the development branch - so
as to incorporate it into a PTB for testing. It is only a couple of lines
and should not interfere with the operation of the code as it only
introduces an extra line of text into the debug output each time a profile
is (re)loaded.

I suspect, but need this to confirm it, that, unlikely as it should be, a
particular package seems to be capable of modifying the execution
environment of the Mudlet application from a profile's Lua interpreter.
This shouldn't be possible, but I am just following up on the pronouncement
of the fictional detective *Sherlock Homes*:
> "You will not apply my precept," he said, shaking his head. "How often
have I said to you that when you have eliminated the impossible, whatever
remains, however improbable, must be the truth?"
>     -- Authur Conan Doyle, *The Sign of the Four*, ch. 6 (1890)

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>