Fix: make Mudlet go full-screen on start-up on small screen devices
#### Brief overview of PR changes/additions
As mentioned in https://github.com/Mudlet/Mudlet/pull/7749#issuecomment-2688604986
this "fixes" the current code-base to work as intended when the option to
play Mudlet on a device with a small screen is enabled. It means that
Mudlet always starts in "Full-screen" mode (without a title-bar) as well as
enabling the display of a main toolbar button (and now a main menu item)
to control the full-screen mode.

The code has also been revised so that enabling/disabling the full screen
mode availability takes effect immediately when the preferences are saved
rather than having to restart Mudlet for this to happen.

#### Motivation for adding to Mudlet
Improving the UX and also making the functionality work as originally
intended.

#### Other info (issues closed, discussion etc)
This WILL clash with #7749 which intended - amongst other things to remove
the control of the Full-screen mode availability completely. That PR will
have to be revised to accommodate this "fix" - the defect having snuck
into #7435.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>