Fix: update map colours when changed in preferences
#### Brief overview of PR changes/additions
Several fixes related to the predefined/preset colours shown in the
preferences and mapped to the Map room environments/colours 257 to 272:
* changing the colours in the preferences immediately changes them in the
map.
* changing the colours via the Lua `setCustomEnvColor(...)` also changes
them immediately - and updates the stored values for those particular
indexes in the custom room colours.
* since the Lua API exposes/allows for an alpha (transparency) the
UI in the preferences now also allows it to be set there - this means the
alternative presentation of the colour superimposed on a black/white
checker-board is used for those elements rather than the whole button being
coloured as is used for permanently 100% opaque colours.
* the above change has also been applied to the map info background as that
already uses transparency but it wasn't available in the settings.

#### Motivation for adding to Mudlet
Makes the UX match the facilities offered.

#### Other info (issues closed, discussion etc)
The original issue was pointed out by @ZookaOnGit in our Discord
**#mudlet-development** channel as of 2025/03/03#06:10
https://www.discord.com/channels/283581582550237184/283582439002210305/1346001658051694686

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>