Improve: add means to close-by-mouse Mudlet when it is full-screen
#### Brief overview of PR changes/additions
Adds a "Close Mudlet" item to the main menu and to the drop down button
of the "Connect" main toolbar button so that it is possible to close Mudlet
using a mouse when it is running in Full-Screen mode - when there won't be
a title bar with a close button on it to use.

#### Motivation for adding to Mudlet
Improved UX when used in the Full-screen case - and given recent PRs like
#7773 or my alternative #7753 which makes the prior code work "as
expected" then this will likely be more often the case.

#### Other info (issues closed, discussion etc)
Whilst developing this I assigned the `QAction::QuitRole` to the menu
item and also the special ones that should have been on a couple of others.
This only has effect on MacOS but causes them to be handled specially when
moved onto the "Application menu" on that platform.

I also experimented with adding a menu item for the *built-in* "About Qt"
dialogue which reports something about the Qt framework generally and the
version actually being used. However I discovered that it is implemented as
a modal dialogue that freezes the event loop, at least on GNU/Linux which
is not suitable for Mudlet (unless we only allow it to be shown if there
is not profiles loaded). Given that we now display the Qt version, (both
the build and run-time ones if they differ) in our own "About Mudlet"
dialogue then this did not seem a useful addition after all.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>