Fix: combine the handling of maximised and full-screen states
It seems that they were interfering with each other!

The previous option `"enableFullScreenMode"` has been removed as it is not
applicable to what is now stored - it will be replaced by `"fullScreen"`.

As there is no need to only create (or in the case of #7753 destroy) the
full-screen controls as per a user setting the code can be simplified and
done in the main `(void) mudlet::init()` method alongside the other
controls.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>