Fix: get Mudlet compiling with Qt 6.9
#### Brief overview of PR changes/additions
Revised codebase to compile with latest Qt version which modified the QChar
constructors such that some casts are needed now to keep things working.
This also revealed a bug in:
`(void) TRoom::auditExits(const QHash<int, int>)` where there was a missing
`10` as a number base argument which was silently being mishandled in prior
versions of Qt.

#### Motivation for adding to Mudlet
Immediately needed for Windows x64 builds but will eventually be needed for
all builds to compile.

#### Other info (issues closed, discussion etc)
Has only come to notice in the last 24 hours but is **HIGH** priority/
urgent.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>