Fix: switch to linuxdeploy
#### Brief overview of PR changes/additions
Revises the Linux CI/CB process to use [`linuxdeploy`](https://github.com/linuxdeploy/linuxdeploy)
instead of `linuxdeployqt`.

#### Motivation for adding to Mudlet
`linuxdeployqt` won't work next week as GitHub are removing the Ubuntu
20.04 runner and that tool will not work with the later `22.04` one until
the maintainer updates it to do so (the CRT is too new).

#### Other info (issues closed, discussion etc)
The maintainer of linuxdeployqt is refusing to updating to Ubuntu 22.04
until it officially reaches EOL (in a month or so) but GitHub will not
provide Ubuntu 20.04 from 2025/04/15 and has been "browning it out
intermittently for the past few months.

This also switches the Linux run that is used for deployment to Ubuntu
22.04 it remains to be seen whether this is still usable for the oldest
versions of the Linux OS which we provide AppImages for.

Some variables in the `bash` scripts used for Linux CI that were not have
been made SCREAMING_SNAKE_CASE so that they can be picked out better.

Some issues that [ShellCheck](https://www.shellcheck.net/) highlighted have been "fixed" as well.

THIS IS A DRAFT CURRENTLY AS IT USES A PR FROM THE **INSTALLERS** REPO.
It can not be merged until that PR is debugged and merged into that
repository.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>