Infrastructure: update Linux runner to Ubuntu 22.04
#### Brief overview of PR changes/additions
Change the Linux Ci/CB runner used to make the AppImage for that OS from
the Ubuntu 20.04 LTS one to the 22.04 one.

#### Motivation for adding to Mudlet
This is unavoidable because the older one is being removed tomorrow.

#### Other info (issues closed, discussion etc)
Because of some intransigence by the maintainer of the `linuxdeployqt`
tool we use to make the AppImage it is not being updated to work with the
22.04 one until that reached its official EOL next month. This seems to
mean that the AppImages produced in the newer runner will NOT work on
systems with the older `libc` that was used in the 20.04 version and others
such as Debian 11 ("Bullseye") and Devuan 4 ("Chimaera").

I have been trying to find a solution but have not got a workable one yet.

(See: https://github.com/Mudlet/Mudlet/pull/7806 and
https://github.com/Mudlet/installers/pull/123).

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>