Infrastructure: fix C++20 and Qt warnings
#### Brief overview of PR changes/additions
* Fix `HostManager::Iter::operator!=` and `HostManager::Iter::operator==`
so the comparison is between equally `const`-ness items.
* Fix `(slot) T2DMap::slot_setRoomProperties(...)` so that one argument
is handled correctly even if it isn't initialised (or used) - by changing
it from a `bool` to a `std::optional<bool>`.
* Change `(void) T2DMap::mouseMoveEvent(QMouseEvent*)` so that it uses
`(QPointF) QMouseEvent::position()` instead of the deprecated in Qt 6
`(QPointF) QMouseEvent::localPos()`.
* Remove some unused, explicit `this` captures in some lambda functions
where they are not used.
* Switch to the **static** `(QStringList) QFontDatabase::families(...)`
from the deprecated in Qt 6 non-static one.
* Revise some `QDateTime` class usages to avoid warnings for things that
have changed in Qt 6.9.0

#### Motivation for adding to Mudlet
Less warnings during the build process.

#### Other info (issues closed, discussion etc)
There are still five warnings for the third party edbee-lib editor widget
involving "arithmetic between different enumeration types" ...
"is deprecated [-Wdeprecated-enum-enumconversion]".

**There is a temporary(?) function added to the `TLuaInterpreter` Lua API
called `dateStamp()` which returns a ISO8601 format timestamp WITH a
timezone offset display at the end. This is to test some code I had to
put together to generate the ones used in the `dlgPackageExporter` class
and it currently produces TWO strings - one using the obsoleted class
methods and the second my attempt to reproduce it. The strings should be
the same as a test. If this is useful I could retain it and just leave the
second (replacement) method in place for future use by others.**

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>