Fix: revert back to luasql-sqlite3 2.6.1 on at least Windows
#### Brief overview of PR changes/additions
Ensure the Windows CI/CB builds are done with the luasql-sqlite3 version
**2.6.1** rather than the **2.7.0** one released (partially) yesterday as
it seems to be breaking Lua database tests that were working the day before
it was release.

#### Motivation for adding to Mudlet
To get Windows builds completing - and passing the Lua tests they did
before 2025/04/23.

#### Other info (issues closed, discussion etc)
This only acts on the Windows builds - the GNU/Linux and MacOS ones would
have to be addressed in the "

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>